Refactor FiltersBox for leaner focus, include in FiltersBoxSearchable
